refactor: update imports from program-section-template-validation to section-template-validation across multiple files
