+++ src/components/common/section-templates/faq-program-section/FaqProgramSection.tsx
@@ -1,4 +1,4 @@
-import { useId, useState, useCallback, useEffect, useRef, useMemo } from 'react';
+﻿import { useId, useState, useCallback, useEffect, useRef, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import cn from 'classnames';
 import { FaqCard } from '@/components/public/faq-section/faq-card/FaqCard';
@@ -11,7 +11,7 @@ import { PublishedFaqQuestion } from '@/types/public/faq-section';
 import { FaqSectionQuestionDto } from '@/types/common/program-sections';
 import { SectionTemplate, SectionMode } from '@/types/common/sections';
 import { ContentType } from '@/types/common/section-contents';
-import { getSectionTemplateMaxLength } from '@/utils/functions/program-section-template-validation/programSectionTemplateValidation';
+import { getSectionTemplateMaxLength } from '@/utils/functions/section-template-validation/sectionTemplateValidation';
 import { getTrimmedInputText } from '@/utils/functions/formatters/text-formatters';
 import { PROGRAM_SECTION_VALIDATION_FUNCTIONS } from '@/validation/admin/program-schema/program-schema';
 import { ReactComponent as PlusIcon } from '@/assets/icons/plus.svg';
