+++ src/components/common/section-templates/shared/title-description-section/TitleDescriptionSection.tsx
@@ -1,4 +1,4 @@
-import styles from './TitleDescriptionSection.module.scss';
+﻿import styles from './TitleDescriptionSection.module.scss';
 import viewStyles from './ViewTitleDescriptionSection.module.scss';
 import cn from 'classnames';
 import { useId } from 'react';
@@ -8,7 +8,7 @@ import { PROGRAMS_TEXT } from '@/const/admin/programs';
 import { COMMON_TEXT_ADMIN } from '@/const/admin/common';
 import { useProgramSectionValidation } from '@/hooks/admin/use-program-section-validation';
 import { getTrimmedInputText } from '@/utils/functions/formatters/text-formatters';
-import { getSectionTemplateMaxLength } from '@/utils/functions/program-section-template-validation/programSectionTemplateValidation';
+import { getSectionTemplateMaxLength } from '@/utils/functions/section-template-validation/sectionTemplateValidation';
 import { SectionMode, SectionTemplate } from '@/types/common/sections';
 import { ContentType } from '@/types/common/section-contents';
