+++ src/utils/functions/section-template-validation/sectionTemplateValidation.ts
